src/requirements.txt
asgiref==3.5.2
Django==4.1.10
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.4
gunicorn==21.2.0